current commit message: continuing switch to hibernate spatial and postgresql in weather related modules\; using pgalise-it project for shared test methods\; switching ejb lookups (which don't seem to work tests outside ear) -> workaround pending\; abstraction of ServiceData and StationData of weathercollector and weather-impl in weather\; removing final modifiers from classes (no valid JPA classes)\;Replaced Weather with StationData because they contain the same information\; hierarchy for StationData and ServiceData\; renaming ServiceWeather to TimeSensitiveServiceDataForecast (contains the same information except Comparable implementation)\; adding getTemperatureUnit method to WeatherService\; removed all redundant classes in weather related modules and made maximal abstraction\; tests don't run yet\;

+++ simulation/de.pgalise.simulation.weather/src/main/java/de/pgalise/simulation/weather/parameter/WindDirection.java
@@ -44,7 +44,7 @@ public final class WindDirection extends WeatherParameterBase {
 	@SuppressWarnings("unchecked")
 	@Override
 	public <T extends Number> T getValue(long time) throws NoWeatherDataFoundException {
-		Integer value = this.getWeather(time).getWindDirection();
+		Float value = this.getWeather(time).getWindDirection();
 		return (T) value;
 	}
